Je vais lancer le programme avant d'aller dormir... Stream04 toujours en travaux ;-(

+++ project-nbaudoux-mdevers-womeire/project-nbaudoux-mdevers-womeire/InputStream01.h
@@ -6,7 +6,7 @@ class InputStream01 : public InputStream
 public:
 	InputStream01();
 	~InputStream01();
-	void open(string, int);
+	void open(string, size_t);
 	int32_t* read_next();
 	bool end_of_stream();
 	void reset();
