+++ project-nbaudoux-mdevers-womeire/project-nbaudoux-mdevers-womeire/InputStream02.h
@@ -6,7 +6,7 @@ class InputStream02 : public InputStream
 public:
 	InputStream02();
 	~InputStream02();
-	void open(string, int);
+	void open(string, size_t);
 	int32_t* read_next();
 	bool end_of_stream();
 	void reset();
