+++ project-nbaudoux-mdevers-womeire/project-nbaudoux-mdevers-womeire/InputStream03.h
@@ -6,7 +6,7 @@ class InputStream03 : public InputStream
 public:
 	InputStream03();
 	~InputStream03();
-	void open(string, int);
+	void open(string, size_t);
 	int32_t* read_next();
 	bool end_of_stream();
 	void reset();
