+++ project-nbaudoux-mdevers-womeire/project-nbaudoux-mdevers-womeire/OutputStream01.h
@@ -6,7 +6,7 @@ class OutputStream01 : public OutputStream
 public:
 	OutputStream01();
 	~OutputStream01();
-	void create(string, int);
+	void create(string, size_t);
 	void write(int32_t*);
 	void close();
 private:
